requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.4.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fire=0.6.0=pypi_0
fonttools=4.49.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
grpcio=1.62.0=pypi_0
h5py=3.10.0=pypi_0
hdf5=1.12.1=h2b7332f_3
hydra-core=1.3.2=pypi_0
idna=3.6=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.3=pyhd33586a_0
ipython=8.22.2=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.1=pypi_0
krb5=1.20.1=h143b758_1
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libcurl=8.9.1=h251f7ec_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.2.0=h807b86a_5
libnghttp2=1.57.0=h2d74bed_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.10.1=pypi_0
llvmlite=0.43.0=pypi_0
loguru=0.7.2=pypi_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.2.1=hd590300_1
packaging=23.2=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
platformdirs=4.2.0=pyhd8ed1ab_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=py310h2372a71_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydeprecate=0.3.1=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python_abi=3.10=2_cp310
pytorch-lightning=1.5.2=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py310h6a678d5_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
sentry-sdk=1.40.5=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.3.0=pypi_0
tifffile=2024.2.12=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.2.0=pypi_0
torchmetrics=1.3.1=pypi_0
torchvision=0.17.0=pypi_0
tornado=6.4=py310h2372a71_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pyhd8ed1ab_0
triton=2.2.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
wandb=0.16.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
